+++ src/main/resources/data.sql
@@ -1,9 +1,3 @@
--- 플러스 친구 개념으로 초기 유저를 넣는다.
--- insert into users(username, password, role)
---     values ('kent beck', 'test password', 'PLUS'),
---            ('mario', 'test password', 'PLUS'),
---            ('119 pc', 'test password', 'PLUS');
-
 insert into users(username, password, role)
 values ('한사랑 이비인후과', 'test password', 'PLUS'),
        ('성지폰', 'test password', 'PLUS'),
